Make a note that Sandbox is sloooooowwwwww (#385)
Based on customer feedback

> Test purchases are slow (like 8 seconds per step) on the testing
devices. This has been discussed in public forums as being Apple's fault
and only happening in development mode and not in production, but there
are confusing comments sometimes and I'm concerned the issue will
persist. It'd be nice to have this behavior documented somewhere if it
isn't.

## Motivation / Description

## Changes introduced

## Linear ticket (if any)

## Additional comments