Paywalls: Add limitations when using developer determined offers (#451)
## Motivation / Description
Currently, developer determined offers don't work great with paywalls
since developers can't choose the subscripiton option they want to
apply. So if a developer uses these offers, the developer determined
offer will always be available, so end users could always have access to
them. This adds a section to the docs mentioning this limitation and
suggesting using the `rc-ignore-offer` for now to be able to avoid using
those offers in paywalls.