Remove emojis and use category slug for folder names (#519)
Using emojis in the folder names imply some issues like these ones
reported by @joshdholtz

## Issue 1

I moved some code blocks into a new category (the ones with the emoji in
the path) and I ran into a weird issue in the sync PR where the code
block was removed from the rendered doc


https://github.com/RevenueCat/revenuecat-docs/pull/346/commits/e03ad1eaf4f7d8a03e59cbf8a8bb971291bf8e99

It turned out, there was a slight unicode issue in the code block path
(not really sure how). The solution was to copy and paste the emoji from
the path (instead of manually typing the emoji like I did)

https://github.com/RevenueCat/revenuecat-docs/pull/346/files

## Issue 2

It is difficult to use git on the CLI with the emoji paths because they
are seen as unicode

----

Apart from that, when creating a new document not yet in Readme.com, the
generated markdown file doesn't have the category slug in the headers.
The solution right now is modifying the rendered file and adding the
category slug manually. If we use the category slug as folder name, we
can directly infer the category slug and add it to the header in the
files in the rendered docs folder.

An alternative is to not use category slugs and just remove the emoji.
Then create a `.slug` file in each folder that has the category slug in
it.