Update Stripe docs for when to send purchases (#244)
## Motivation / Description
From a Zendesk ticket, it was found that when Stripe customer support
gives a subscription, then the checkout event is not triggered so this
will be missed. To fix this, developers should listen to and send Stripe
receipts at both the checkout and subscription created events. No harm
in doing this as we would be able to tell if it is a duplicate and would
not create two events.

## Changes introduced
Added to listen to subscription created events as well.

## Jira ticket (if any)
https://revenuecat.zendesk.com/agent/tickets/27974

---------

Co-authored-by: Cody Kerns <44073103+codykerns@users.noreply.github.com>
Co-authored-by: Cesar de la Vega <cesarvegaro@gmail.com>