Change "current" Offering to "default" Offering when referencing a Project (#542)
## Motivation / Description

Full background is
[here](https://revenuecat.slack.com/archives/CFU1JG9U6/p1700595848075709).
TL;DR version is:
1. When we release Targeting, the term current Offering will become even
less appropriate than it already is, so we're taking the opportunity to
change it default Offering when referencing "a Project's default
Offering."
2. In the context of getting Offerings, though, we'll continue to
describe the Offering that should be served for a customer as **their**
current Offering. So no SDK/API changes are needed, and in the docs and
Dashboard anytime we are referring to what an individual customer will
be served (e.g. in the Customer Profile) we will continue to call that
their current Offering.

## Changes introduced

I went through all of the places in our docs that need to be updated,
but definitely need a QA to make sure:
1. My changes make sense, and
2. I didn't miss any

## Linear ticket (if any)


https://linear.app/revenuecat/issue/PWL-426/update-docs-for-the-current-to-default-terminology-change

## Additional comments

---------

Co-authored-by: Josh Holtz <me@joshholtz.com>